requirements.txt
Django==2.2.1
xlwt==1.3.0
xlrd==1.2.0
pycrypto==2.6.1
django_fernet_fields==0.6
pymysql==0.9.3